requirements.txt
# Core dependencies
openai>=1.50.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0

# Job scraping
python-jobspy>=1.1.80

# Database
sqlalchemy>=2.0.0

# Data validation
pydantic>=2.0.0

# Configuration
pyyaml>=6.0.0

# CLI
click>=8.1.0
rich>=13.0.0

# Utilities
python-dateutil>=2.8.0

# Notifications
python-telegram-bot>=21.0.0
